+++ src/test/scala-2/chiselTests/simulator/scalatest/ChiselSimSpec.scala
@@ -4,12 +4,13 @@ package chiselTests.simulator.scalatest
 
 import chisel3._
 import chisel3.simulator.PeekPokeAPI.FailedExpectationException
-import chisel3.simulator.{ChiselSettings, ChiselSim, MacroText}
+import chisel3.simulator.{ChiselSettings, ChiselSim, HasSimulator, MacroText}
 import chisel3.testing.HasTestingDirectory
 import chisel3.testing.scalatest.{FileCheck, TestingDirectory}
 import java.nio.file.FileSystems
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import scala.reflect.io.Directory
 
 class ChiselSimSpec extends AnyFunSpec with Matchers with ChiselSim with FileCheck with TestingDirectory {
 
@@ -38,6 +39,25 @@ class ChiselSimSpec extends AnyFunSpec with Matchers with ChiselSim with FileChe
           foo.a.poke(false.B)
           foo.b.expect(false.B)
         }
+      }
+    }
+
+    it("should error if an expect fails") {
+      val message = intercept[Exception] {
+        simulate {
+          new Module {
+            val a = IO(Output(Bool()))
+            a :<= false.B
+          }
+        } { _.a.expect(true.B) }
+      }.getMessage
+      message.fileCheck() {
+        """|CHECK:      Failed Expectation
+           |CHECK-NEXT: ---
+           |CHECK-NEXT: Observed value: '0'
+           |CHECK-NEXT: Expected value: '1'
+           |CHECK:      ---
+           |""".stripMargin
       }
     }
 
@@ -120,6 +140,38 @@ class ChiselSimSpec extends AnyFunSpec with Matchers with ChiselSim with FileChe
              |CHECK-NEXT: '+define+STOP_COND=!svsimTestbench.c'
              |""".stripMargin
         )
+    }
+
+    it("should allow for a user to customize the build directory") {
+      class Foo extends Module {
+        stop()
+      }
+
+      /** An implementation that always writes to the subdirectory "test_run_dir/<class-name>/foo/" */
+      implicit val fooDirectory = new HasTestingDirectory {
+        override def getDirectory =
+          FileSystems.getDefault().getPath("test_run_dir", "foo")
+      }
+
+      val directory = Directory(FileSystems.getDefault().getPath("test_run_dir", "foo").toFile())
+      directory.deleteRecursively()
+
+      simulate(new Foo()){ _ => }(hasSimulator = implicitly[HasSimulator], testingDirectory = fooDirectory)
+
+      info(s"found expected directory: '$directory'")
+      assert(directory.exists)
+      assert(directory.isDirectory)
+
+      val allFiles = directory.deepFiles.toSeq.map(_.toString).toSet
+      for (
+        file <- Seq(
+          "test_run_dir/foo/workdir-verilator/Makefile",
+          "test_run_dir/foo/primary-sources/Foo.sv"
+        )
+      ) {
+        info(s"found expected file: '$file'")
+        allFiles should contain(file)
+      }
     }
   }
